Support parsing a Sec-WebSocket-Extensions header with multiple elements
This change is mainly for the WebSocket server implementation in Chromium.

As there's only one known extension so far and it's illegal to return
multiple permessage-deflate elements to the client, we don't need this
parser improvement for the client.

R=yhirano,dgozman,eroman
BUG=471646

Review URL: https://codereview.chromium.org/1047623002

Cr-Original-Commit-Position: refs/heads/master@{#323189}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 38ee68c20fe305db2eb4f3db50d3c9cbfad03d45