s/NPN_SPDY_3/NPN_SPDY_4/g in tests.
Since SPDY/3 is deprecated, replace it with HTTP/2 (also known as SPDY/4
internally) in tests.  This is not critical as they are the negotiated protocols
are not actually used in the tests affected by this CL, but it is a prerequisite
to removing NPN_SPDY_3 enum value from the codebase.

BUG=428086

Review URL: https://codereview.chromium.org/1059243003

Cr-Original-Commit-Position: refs/heads/master@{#323759}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6d9b2f470073c2125c1e085e38efa7710d94aca1