CHECK() if Windows non-blocking connect returns 0.
Non-blocking connect() on Windows is documented as always returning
WSAEWOULDBLOCK or another error.

There has been code in place to handle a 0 return value, but we have
no way to test whether it is correct or not.

Replace the code to handle a 0 return with a CHECK() call, to find out
whether a 0 result really can happen or not.

BUG=436634
TEST=net_unittests

Review URL: https://codereview.chromium.org/1055813006

Cr-Original-Commit-Position: refs/heads/master@{#326243}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2d53a9381e88eab543cc4c9fd122b0b1ed4f41d7