Report the connect status of the oldest connection in the socket pool.
This CL implements a combination of the suggestions 1) and 2)
and comment #1 in crbug.com/462808. This reduces the number of
calls to GetLoadState. However the status might be less accurate
as we are far from guaranteed that the oldest connection is the
one which has come the furthest. This should be acceptable though,
as the state is mainly used report activity or at which stage
connections are hung at.

BUG=462808

Review URL: https://codereview.chromium.org/1091793002

Cr-Original-Commit-Position: refs/heads/master@{#326244}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 835c1d6d6b8374ed2e71e04825b7b3c58fd0aa32