Refactor NetLog::LogLevel --> NetLogCaptureMode.
This is both a rename, and a functional change.

Before LogLevel was an enum, and consumers would compare its numeric value, or use it for serialization purposes. Whereas NetLogCaptureMode is a class that encapsulates the various capabilities for a logging level.

BUG=472710
TBR=asanka@chromium.org,sergeyu@chromium.org,yurys@chromium.org,bengr@chromium.org,asargent@chromium.org,zea@chromium.org,scherkus@chromium.org

Review URL: https://codereview.chromium.org/1059843002

Cr-Original-Commit-Position: refs/heads/master@{#326440}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 001c374c5356777641161bb11b8dd792508381ca