Route OCSP stapling through CertVerifier.
This moves the platform-specific OCSP stapling code out of SSLClientSocket,
which enables OCSP stapling on the chimera build. It'll also make it easier to
support OCSP stapling on OS X, where a stapled OCSP response is a property of
the certificate verification (as it should be) and not the certificate.

BUG=430714,479034

Review URL: https://codereview.chromium.org/1081913003

Cr-Original-Commit-Position: refs/heads/master@{#327070}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 15f5713c41825ea900cd6dc4f6de4a6d3fff538d