(Reland) Enable ignoring network interfaces in NetworkChangeNotifierLinux.
NetworkChangeNotifier should not report network changes from interfaces
that are not used to connect to the internet (ie. wifi APs). Adjust constructor
for NetworkChangeNotifierLinux to accept a list of interfaces that are known to
not be used to connect to internet.

R=pauljensen@chromium.org,akuegel@chromium.org,jochen@chromium.org
BUG=469863

Review URL: https://codereview.chromium.org/1128533003

Cr-Original-Commit-Position: refs/heads/master@{#328378}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 09c74490237aeefe5f41c1cad07a7d1e2b384590