(Reland) Enable ignoring network interfaces in NetworkChangeNotifierLinux.
NetworkChangeNotifier should not report network changes from interfaces
that are not used to connect to the internet (ie. wifi APs). Adjust constructor
for NetworkChangeNotifierLinux to accept a list of interfaces that are known to
not be used to connect to internet.

BUG=469863

Review URL: https://codereview.chromium.org/1127303002

Cr-Original-Commit-Position: refs/heads/master@{#328915}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ea1725c8b975e9ad925463bf65ffa6897adcd8df