Remove SpdyStream::GetSSLCertRequestInfo.
SPDY's multiplexing is incompatible with the renego + client auth hack. The old
implementation also assumed that the SpdyStream was still alive which isn't
true. We could attempt save a copy of it, but that doesn't resolve the
fundamental incompatibility of the hack. Instead
https://codereview.chromium.org/1131763002/ disables renegotiation, so this
codepath should be impossible.

BUG=462283

Review URL: https://codereview.chromium.org/1128873008

Cr-Original-Commit-Position: refs/heads/master@{#329493}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b38fe83bd9f34f101e705bff5397c9530b024edd