Disable 0RTT if server and origin have different hosts.
Disable 0RTT if the hostname of the server and origin are different.  Currently
dead code since HttpStreamFactoryImpl::GetAlterantiveServiceFor() refuses to
return an altenative service with hostname different than that of the origin,
but I plan to change this very soon.

While disabling 0RTT in this case is not strictly necessary, it enables us to
move faster with alternative service implementation.  We expect the most
performance gain from pooling to existing connections to different hosts, which
is not hindered by this change.

* Add origin hostname parameter to QuicStreamRequest::Request() method.
* Add server_and_origin_have_same_host parameter to QuicStreamFactory::Create()
  and CreateAuxiliaryJob() methods, and QuicStreamFactory::Job constructor.
* Add server_and_origin_have_same_host_ member to QuicSteamFactory::Job.
* Driveby: rename was_alternate_protocol_recently_broken_ member to
  was_alternative_service_recently_broken_.

BUG=474217

Review URL: https://codereview.chromium.org/1140153003

Cr-Original-Commit-Position: refs/heads/master@{#330402}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68d401ddf26908913f659ef7d4450df351ae33cb