Prevent unlimited re-sends of failed partial requests.
If we receive a 401 on a network request, after we've already read some
data successfully, then we will retry the request.  This change introduces
a fixed limit (currently 1) on the number of consecutive retries.

Previously, we would retry indefinitely if we continue to get 401s.  We
could hit this case in the case of credentials that expire part way through
a read.

BUG=474835

Review URL: https://codereview.chromium.org/1103563002

Cr-Original-Commit-Position: refs/heads/master@{#331965}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1f59bb3f4c02afbc538ca78e6d1f5d1b82a75b5e