Fixing incorrect certificate requirement for QUIC proxy.
HttpStreamFactoryImpl::Job::using_quic_ can be set not only for QUIC alternative
Jobs, but also when using a QUIC proxy.  In this case, however, the (proxy)
server should not be required to exhibit a certificate for the origin.  This CL
sets this straight.

BUG=492458

Review URL: https://codereview.chromium.org/1174643002

Cr-Original-Commit-Position: refs/heads/master@{#333960}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 313ba9c225e67d24c0e661580f7129d4304f5e5d