Add served cert chain to SSLInfo
Add a field to SSLInfo to contain the chain as received by the client,
which can be different than the SSLInfo::cert containing the chain that
was verified. The new |served_cert| field is not persisted to the cache
for an HttpResponseInfo and thus can be null.

BUG=498945

Review URL: https://codereview.chromium.org/1174073003

Cr-Original-Commit-Position: refs/heads/master@{#334292}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 03d644fe40b663dd42397649848484e87062afd6