Disable the TLS 1.0 version fallback.
Re-enable the SSLFallbackVersionMin admin policy, this time documented for the
TLS 1.0 fallback.

As with when the SSL 3.0 was removed, we for now will continue to attempt
fallbacks beyond the minimum version. But, if they succeed, the connection is
shut off and a dedicated error code is returned. This slightly tweaks the
string for that error code since this isn't in response to, e.g. POODLE, unlike
with SSL 3.0.

(Although, being able to securely negotiate > TLS 1.0 means we don't have to
worry about BEAST, in case the server's not doing record-splitting.)

This does not yet remove the TLS 1.1 fallback leg.

BUG=498998

Review URL: https://codereview.chromium.org/1173193004

Cr-Original-Commit-Position: refs/heads/master@{#334430}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b127ca8a1aca94228508d74ec84e5bed4be4f2da