Use a per-connection writer in the QuicTimeWaitListManager.
Fixes bug reported on proto-quic mailing list:
https://groups.google.com/a/chromium.org/forum/#!topic/proto-quic/ziwUgZbV2Wk

To reproduce:
1. Run the toy server
2. Connect to the server with chrome
3. Quit toy server and restart (don't quit chrome)
4. Reload chrome page
5. Fatal check fail

Previously the QuicTimeWaitListManager was calling the WritePacket method of the QuicSimpleServerPacketWriter, which the comments make clear should not be done. Wrapping this in a QuicSimplePerConnectionPacketWriter (the result of calling the Create method of the factory) means the packet is written out correctly, ultimately using WritePacketWithCallback.

BUG=

Review URL: https://codereview.chromium.org/1192953004

Cr-Original-Commit-Position: refs/heads/master@{#335334}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 54d59d5218042438aba7f607481d1e28dce895e8