Set alternative service even if it is broken.
Modify HttpServerPropertiesImpl::SetAlternativeService() not to ignore a broken
alternative service.  This is consistent with the current behavior of keeping
broken alternative services in the mapping, and of refusing to use the
alternative service of a canonical host if the origin has an alternative
service, even if it is broken.  Both of these show that broken alternative
services have a significance and thus should not be ignored as input of
SetAlternativeService.

BUG=392575

Review URL: https://codereview.chromium.org/1205633002

Cr-Original-Commit-Position: refs/heads/master@{#336145}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eb2e6601f1851335bbef50127805d5c9ed4f92db