Ignore certificate transparency by default.
Current behavior is to remove the EV flag if the ct enforcer is not
set by the embedder. This somewhat unexpectedly removes EV for all
servers, if the embedder has not taken explicit actions to turn
on CT.

CT should either be ignored when policy enforcer is not present,
or there should be asserts warning about the missing EV-white
list. This patch chooses the first option as CT is not yet a
security requirement, and I don't think it's correct to demand
CT support quite yet.

Note that this change has no effect on Chrome's default behavior.

BUG=NONE

Review URL: https://codereview.chromium.org/1211423002

Cr-Original-Commit-Position: refs/heads/master@{#336539}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2553ff05b802a94ef281e647874d37941eefd154