Fix some clang warnings with -Wmissing-braces in net.
Clang warns if there are missing braces around a subobject
initializer. The most common idiom that triggers this is:
  STRUCT s = {0};
if the first field of STRUCT is itself a struct. This can
be more simply written as:
  STRUCT s = {};
which also prevents the warning from firing.

BUG=505297

Review URL: https://codereview.chromium.org/1217403002

Cr-Original-Commit-Position: refs/heads/master@{#337261}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e4e094f274a21a97909fe3a11e79b4d48eabd7c4