mandoline filesystem: Reland "Save cookie data to the mojo:filesystem."
In addition to saving the cookies to the file system service (per the
previous patch), we also modify the filesystem service to take a
client. When the shell connection to the file system service is broken,
instead of quitting immediately, we broadcast a shutdown notification to
all of our clients, and wait for them to close the directory objects.

This solves the timing problems during shutdown.

BUG=493311
TBR=shess@chromium.org
R=jam@chromium.org,erikwright@chromium.org
First Review URL: https://codereview.chromium.org/1179413010

Review URL: https://codereview.chromium.org/1231493002

Cr-Original-Commit-Position: refs/heads/master@{#338888}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e3ec7c45bd64ef826fd3940c8a2302f5b55edc45