Move cookie source display logic to CookiesTreeModel
Previously, CanonicalCookie::Source() returned the source of the cookie
as needed by CookiesTreeModel -- that is, with the port stripped and
scheme changed to http. This CL moves this CookiesTreeModel-specific
logic to CookiesTreeModel, so that CanonicalCookie::Source() represents
the actual origin that set the cookie.

This CL is to be followed up with
https://codereview.chromium.org/1303593002/, which aims to measure the
frequency of non-secure origins setting Secure cookies.

BUG=522261
TBR=miguelg@chromium.org

Review URL: https://codereview.chromium.org/1297023004

Cr-Original-Commit-Position: refs/heads/master@{#344866}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bd99e6be869fd6caff8a0f676d1357dacdbc7532