Record scheme of URLs that set Secure cookies
This CL histograms the type of scheme (cryptographic or not
cryptographic) for URLs that set or overwrite Secure cookies. The goal
is to measure whether we can remove the ability for http:// URLs to set
Secure cookies.

BUG=522261

Review URL: https://codereview.chromium.org/1303593002

Cr-Original-Commit-Position: refs/heads/master@{#344907}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7feb65c2b8098104c34a1528556553b535066659