Throw out same-host HTTPS HPKP report URIs
If A.com has a pinning violation which triggers a report sent to A.com,
Chrome will go into a loop, because sending the report will itself incur
a pinning violation, which will trigger a report, and so on.

Therefore this CL throws out HTTPS report URIs that use the same host as
the corresponding pin.

BUG=516910

Review URL: https://codereview.chromium.org/1301273002

Cr-Original-Commit-Position: refs/heads/master@{#346722}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3b27c57e03c7938d17400be95058e709c198e387