Expose currently-connected socket address in URLRequest
URLRequest has a GetSocketAddress method, but it returns the socket
address stored in HttpResponseInfo, which is only updated after a
response is received (and is pre-populated from the cache when
validating a cached resource). This means that it can fail to return
an address when the net stack is actually connected, or it can return a
stale address during cache revalidation.

Plumb through a new GetRemoteEndpoint method that specifically
gets the address straight from the socket layer (with a stop to cache
it in the HttpNetworkTransaction).

(This is intended for Domain Reliability, but can be used elsewhere.)

BUG=480565

Review URL: https://codereview.chromium.org/1270093005

Cr-Original-Commit-Position: refs/heads/master@{#351383}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d9dbc65205e077bb161573a51995710b767fc061