[Android] Introduce RegistrationPolicy for NetworkChangeNotifier.
Currently the NetworkChangeNotifier.setAutoDetectConnectivityState(true) uses
ApplicationState to determine when to listen to network changes. However the
ApplicationStatus class does not work as such in the context of WebView. To make
the NetworkChangeNotifier more flexible this patch introduces a RegistrationPolicy
class which can be passed to the overloaded setAutoDetectConnectivityState method.

We make sure to keep the overall API compatible, such that existing users of the
NetworkChangeNotifier do not have to provide a policy, which is set behind the
screens instead. To this end this patch includes two concrete implementations of
the RegistrationPolicy.

BUG=520088

Review URL: https://codereview.chromium.org/1358163004

Cr-Original-Commit-Position: refs/heads/master@{#353071}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 97997f341e5e7205b5311a5f1ed66f8cd0d66454