Remove HttpResponseHeaders::GetParsedHttpVersion().
The difference between GetParsedHttpVersion() and GetHttpVersion() is subtle and
obscure.  (The former is not clamped, the latter is.)  This cleanup CL removes
the GetParsedHttpVersion() method and parsed_http_version_ member (using a local
variable instead in ParseStatusLine()).

This changes behavior slightly for servers that advertise HTTP/0.9 but have
headers, which are now treated as HTTP/1.0 in
HttpNetworkTransaction::DoReadHeadersComplete().

BUG=

Review URL: https://codereview.chromium.org/1408713003

Cr-Original-Commit-Position: refs/heads/master@{#354299}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: be0f6af46753add1f53a31243226db5927e7d612