[Cronet] Switch from InstrumentationTest to junit
Instrumentation tests are meant for testing apps, where as junit tests
are better for unit testing, which is what the cronet tests do.
This change gets rid of the restriction of setting up the tests via
an Intent.  It removes a fair bit of code and paves the way for
removing more, including JSON serializing of CronetEngine.Builder.

I think this makes testing faster...because I ran into two race
conditions in tests that I also fixed:
1. Starting the CronetTestFramework with a URL did not wait for the
   request to finish.
2. NetworkChangeNotifier initialization test did not wait for the
   NetworkChangeNotifier to actually initialize.

For now the test continues to be launched as an instrumentation test
because this ensures lighttpd is launched, which isn't the case for
junit tests.  All tests run and pass being launched as an
instrumentation test.

BUG=544976

Review URL: https://codereview.chromium.org/1417973002

Cr-Original-Commit-Position: refs/heads/master@{#355770}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 373bed9bc93f00af4244808cff41c355de268ed7