Fix a pair of use-after-destructs in Mojo proxy resolver unittests.
A ProxyResolverMojo has a callback that's run on completion. In these
tests, that callback is created and passed from the resolver factory and
points to a factory member.  However, the factory destroys that member
before destroying the resolver, which results in a
use-after-destruction, which is undefined behaviour in C++ (according
to the spec).

Review URL: https://codereview.chromium.org/1429513005

Cr-Original-Commit-Position: refs/heads/master@{#356232}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 41826031a78af14dc53ab525f01aa0613bef44f9