Fix obscure bug with URLRequestHttpJob cancellation.
If URLRequestHttpJob::StartTransactionInternal failed
synchronously, transctaion_ would be NULL and a task to pass
the error to URLRequestHttpJob::OnStartCompleted would be posted.
If the request is cancelled before this task is run, then the
cancellation would write URLRequestStatus and then call
URLRequestHttpJob::Kill, which would do nothing, since the
transaction_ was NULL.  OnStartCompleted would then be run,
overwrite the cancelled status with an error, and then treat
it as a failure.

While not a major bug, overwriting the cancelled status with
the error is weird, and breaks a CHECK aimed at catching real
crashers.

BUG=508900

Review URL: https://codereview.chromium.org/1414823005

Cr-Original-Commit-Position: refs/heads/master@{#357421}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2281f376ebc04e8512453e12adac7e497b3dbf3c